Reorganize module structre
