--- frontend/src/index.js
@@ -1,1 +1,0 @@
-import App from "./components/App";
