+++ frontend/src/pages/Dashboard.js
@@ -2,9 +2,8 @@ import Button from "@material-ui/core/Button";
 import React from "react";
 import {makeStyles} from "@material-ui/core/styles";
 import {Link} from "react-router-dom";
-import {Breadcrumbs} from "@material-ui/core";
 import Typography from "@material-ui/core/Typography";
-import {grey} from "@material-ui/core/colors";
+import AironeBreadcrumbs from "../components/AironeBreadcrumbs";
 
 const useStyles = makeStyles((theme) => ({
     button: {
@@ -14,11 +13,6 @@ const useStyles = makeStyles((theme) => ({
         display: "flex",
         justifyContent: "center",
     },
-    breadcrumbs: {
-        padding: theme.spacing(1),
-        marginBottom: theme.spacing(1),
-        backgroundColor: grey[300],
-    },
 }));
 
 export default function Dashboard(props) {
@@ -26,9 +20,9 @@ export default function Dashboard(props) {
 
     return (
         <div>
-            <Breadcrumbs aria-label="breadcrumb" className={classes.breadcrumbs}>
+            <AironeBreadcrumbs>
                 <Typography color="textPrimary">Top</Typography>
-            </Breadcrumbs>
+            </AironeBreadcrumbs>
 
             <div className={classes.links}>
                 <Button
