+++ frontend/src/pages/Job.js
@@ -1,8 +1,6 @@
 import {makeStyles} from "@material-ui/core/styles";
-import {grey} from "@material-ui/core/colors";
 import React, {useEffect, useState} from "react";
 import {
-    Breadcrumbs,
     List,
     ListItem,
     Table,
@@ -16,6 +14,7 @@ import Typography from "@material-ui/core/Typography";
 import {Link} from "react-router-dom";
 import Button from "@material-ui/core/Button";
 import Paper from "@material-ui/core/Paper";
+import AironeBreadcrumbs from "../components/AironeBreadcrumbs";
 
 const useStyles = makeStyles((theme) => ({
     button: {
@@ -23,11 +22,6 @@ const useStyles = makeStyles((theme) => ({
     },
     entityName: {
         margin: theme.spacing(1),
-    },
-    breadcrumbs: {
-        padding: theme.spacing(1),
-        marginBottom: theme.spacing(1),
-        backgroundColor: grey[300],
     },
 }));
 
@@ -52,10 +46,10 @@ export default function Job(props) {
 
     return (
         <div className="container-fluid">
-            <Breadcrumbs aria-label="breadcrumb" className={classes.breadcrumbs}>
+            <AironeBreadcrumbs>
                 <Typography component={Link} to='/new-ui/'>Top</Typography>
                 <Typography color="textPrimary">ジョブ一覧</Typography>
-            </Breadcrumbs>
+            </AironeBreadcrumbs>
 
             <div className="row">
                 <div className="col">
